Added LOCK, UNLOCK, TIMEOUT, and FAIL_EARLY flags to sets; in support of locks

+++ src/include/callbacks.h
@@ -2,7 +2,7 @@
 #define ITCALLBACKSH
 #include "intheory.h"
 
-void slot_changed(long slot, long value);
+void slot_changed(int slot, long value);
 void destroy_cb();
 
 #endif
